Improve: FP8 GEMM throughput on Skylake/Haswell + Granite Rapids E5M2 kernel
E5M2 cast (`cast/{skylake,haswell}.h`):
- Rewrite `nk_e5m2x{16,8}_to_f32x{16,8}_*` as 3-op (cvtepu8 + slli 8 + cvtph_ps).
  E5M2 shares F16 bias, so `byte << 8` is the matching F16 encoding.

Pairwise dot (`dot/{skylake,haswell}.h`):
- E5M2 widened to 64/16-lane multi-chain inline unpack:
  Skylake 3.17-3.94×, Haswell 2.07-5.05× across D=100..768.
- E4M3 unchanged (16-lane single-chain was already at the cast cost limit).

Stateful GEMM (`dot/`, `dots/{skylake,haswell}.h`):
- E5M2: byte-pack + new dtype-specific `nk_dot_e5m2x{64,32}_update_*` with
  two independent FMA chains folding into the single state accumulator.
  Skylake dots/angulars/euclideans 1.42-2.52×, Haswell up to 3.17×.
- E4M3 Skylake: asymmetric F16-pack (A streamed as F32, B pre-cast and
  stored as F16, widened to F32 on load). 50% memory savings for the
  packed B; compute ~0.89-1.16× of baseline (accepted tradeoff).
- E4M3 Haswell: byte-pack at depth=32, neutral 0.95-1.03×.

New Granite Rapids E5M2 GEMM (`dots/graniteamx.h`, `spatials/graniteamx.h`):
- Pack E5M2 → F16 via `byte << 8`, run TDPFP16PS over F16 tiles to F32.
  Beats Sapphire AMX BF16 path on E5M2 inputs (better intermediate
  precision, same throughput). Wired through `dispatch_e5m2.c` ahead
  of Sapphire AMX.

Sapphire AMX (`dots/sapphireamx.h`): keeps icelake LUT cast helpers
(empirical eval showed Genoa-Giesen path regressed; revert preserved).

Spatials (`spatials/skylake.h`): E4M3 normalize_packed packed_value_type
follows dots packed_value_type (f16) for matching norm offset.

Bonus: collapse split string literals across JS/Python error messages
(`javascript/numkong.c`, `python/{each,matrix,tensor}.c`).

ULP: dots/spatials max_ulp ≤ 1 vs scalar reference (threshold 32).

+++ include/numkong/spatials/skylake.h
@@ -97,11 +97,11 @@ nk_define_cross_normalized_symmetric_(euclidean, bf16, skylake, bf16, f32, /*nor
                                       nk_dots_reduce_sumsq_bf16_, nk_load_b128_haswell_, nk_partial_load_b32x4_skylake_,
                                       nk_store_b128_haswell_, nk_partial_store_b32x4_skylake_, 1)
 
-nk_define_cross_normalized_packed_(angular, e4m3, skylake, e4m3, f32, f32, /*norm_value_type=*/f32, f32, nk_b128_vec_t,
+nk_define_cross_normalized_packed_(angular, e4m3, skylake, e4m3, f16, f32, /*norm_value_type=*/f32, f32, nk_b128_vec_t,
                                    nk_dots_packed_e4m3_skylake, nk_angular_through_f32_from_dot_haswell_,
                                    nk_dots_reduce_sumsq_e4m3_, nk_load_b128_haswell_, nk_partial_load_b32x4_skylake_,
                                    nk_store_b128_haswell_, nk_partial_store_b32x4_skylake_, 1)
-nk_define_cross_normalized_packed_(euclidean, e4m3, skylake, e4m3, f32, f32, /*norm_value_type=*/f32, f32,
+nk_define_cross_normalized_packed_(euclidean, e4m3, skylake, e4m3, f16, f32, /*norm_value_type=*/f32, f32,
                                    nk_b128_vec_t, nk_dots_packed_e4m3_skylake,
                                    nk_euclidean_through_f32_from_dot_haswell_, nk_dots_reduce_sumsq_e4m3_,
                                    nk_load_b128_haswell_, nk_partial_load_b32x4_skylake_, nk_store_b128_haswell_,
